[chore][pkg/stanza] Cleanup unquote operator files
