+++ pkg/stanza/operator/transformer/unquote/transformer_test.go
